tree structure update #2
