--- intent_client/src/components/NodeListHeader.tsx
@@ -1,16 +1,0 @@
-import {Col, Row} from "react-bootstrap";
-import React from "react";
-
-function NodeListHeader() {
-
-
-    return (
-        <Row>
-            <Col>
-                <h3 className={'text-center'}>Node list</h3>
-            </Col>
-        </Row>
-    )
-}
-
-export default NodeListHeader
