+++ intent_client/src/features/intentChat/components/Message/index.scss
