+++ intent_client/src/features/intentChat/components/Message/index.types.ts
@@ -1,0 +1,5 @@
+import {IntentMessage} from "../../types";
+
+export interface MessageProps {
+    message: IntentMessage
+}
