+++ intent_client/src/features/intentChat/components/Window/index.scss
