+++ intent_client/src/features/intentChat/components/Window/index.types.ts
@@ -1,0 +1,5 @@
+interface ChatWindowProps {
+    isOpen: boolean,
+    onClose: () => void,
+    title: string
+}
