+++ intent_client/src/features/topology/components/NodeList/NodeList.types.tsx
@@ -1,3 +1,5 @@
 interface ListNodeProps {
     title: string,
 }
+
+interface DraggableNode extends ListNodeProps {}
